+++ lib/src/common/assets.dart
@@ -6,7 +6,7 @@ abstract class YouthAssets {
 class _Images {
   const _Images();
 
-  static const _basePathImage = 'assets/images';
+  static const _basePathImage = 'packages/youth_biomarkers_sdk/assets/images';
 
   final scanFrame = '$_basePathImage/img_scan_frame.png';
 }
@@ -14,7 +14,7 @@ class _Images {
 class _Files {
   const _Files();
 
-  static const _basePathFiles = 'lib/assets/files';
+  static const _basePathFiles = 'packages/youth_biomarkers_sdk/assets/files';
 
   final liqaScript = '$_basePathFiles/liqa_script.txt';
 }