+++ lib/src/common/mapper/youth_result_mapper.dart
@@ -1,9 +1,12 @@
 import 'package:biosensesignal_flutter_sdk/vital_signs/vital_signs_results.dart';
 import 'package:biosensesignal_flutter_sdk/vital_signs/vitals/vital_sign.dart';
 import 'package:youth_biomarkers_sdk/src/biomarker_services/binah_sdk/binah_bio_scheme.dart';
+import 'package:youth_biomarkers_sdk/src/biomarker_services/liqa_sdk/haut_bio_scheme.dart';
 import 'package:youth_biomarkers_sdk/src/common/scheme/youth_bio_scheme.dart';
 import 'package:youth_biomarkers_sdk/src/common/scheme/youth_bio_scheme_item.dart';
 import 'package:youth_biomarkers_sdk/src/common/scheme/youth_bio_type.dart';
+import 'package:youth_biomarkers_sdk/src/data/models/haut_analyze_marker_model.dart';
+import 'package:youth_biomarkers_sdk/src/data/models/haut_analyze_model.dart';
 import 'package:youth_biomarkers_sdk/src/wrapper/youth_data_point.dart';
 
 final class YouthResultMapper {
@@ -39,6 +42,34 @@ final class YouthResultMapper {
       BinahBioScheme.snsIndex => YouthBioType.binahSns,
       BinahBioScheme.stressLevel => YouthBioType.binahStressResponse,
       BinahBioScheme.respirationRate => YouthBioType.binahRespiratoryFitness,
+      _ => YouthBioType.unknown
+    };
+
+    return youthBioScheme.firstWhere((item) => item.type == youthType);
+  }
+
+  static List<YouthDataPoint> handleHautResults(HautAnalyzeModel hautModel) {
+
+    return hautModel.data.map((item) => _handleHautResult(item)).toList();
+  }
+
+  static YouthDataPoint _handleHautResult(HautAnalyzeMarkerModel result) {
+    final schemeItem = _convertHautTypeToYouth(result.type);
+    return YouthDataPoint(
+      type: schemeItem.type.key,
+      name: schemeItem.name,
+      value: result.value,
+      unit: schemeItem.unit,
+    );
+  }
+
+  static YouthBioSchemeItem _convertHautTypeToYouth(String type) {
+    final youthType = switch(type) {
+      HautBioScheme.faceSkinRedness => YouthBioType.faceSkinRedness,
+      HautBioScheme.faceSkinUniformness => YouthBioType.faceSkinUniformness,
+      HautBioScheme.faceSkinHydration => YouthBioType.faceSkinHydration,
+      HautBioScheme.faceSkinPoresHealth => YouthBioType.faceSkinPoresHealth,
+      HautBioScheme.faceSkinPerceivedAge => YouthBioType.faceSkinPerceivedAge,
       _ => YouthBioType.unknown
     };