+++ lib/src/common/scheme/youth_bio_type.dart
@@ -20,7 +20,17 @@ enum YouthBioType {
   binahRecoverAbility("binah_recover_ability"),
   binahStressResponse("stress_response"),
   binahSns("binah_sns"),
-  binahPns("binah_pns");
+  binahPns("binah_pns"),
+  faceSkinRedness("face_skin_redness"),
+  faceSkinUniformness("face_skin_uniformness"),
+  faceSkinHydration("face_skin_hydration"),
+  faceSkinPoresHealth("face_skin_pores_health"),
+  faceSkinAcneIndex("face_skin_acne_index"),
+  faceSkinWrinklesIndex("face_skin_wrinkles_index"),
+  faceSkinPigmentation("face_skin_pigmentation_index"),
+  faceSkinEyeAreaCondition("face_skin_eye_area_condition"),
+  faceSkinPerceivedAge("face_skin_perceived_age"),
+  faceSkinEyeAge("face_skin_eye_age");
 
   final String key;