+++ lib/youth_sdk_exports.dart
@@ -1,12 +1,13 @@
 library;
 
 export 'src/youth_sdk.dart';
+export 'src/common/assets.dart';
 export 'src/common/scheme/youth_bio_type.dart';
 export 'src/common/scheme/youth_bio_measure_type.dart';
 export 'src/controller/youth_video_controller.dart';
-export 'src/controller/youth_selfie_controller.dart' show YouthSelfieController;
-export 'src/biomarker_services/binah_sdk/view/youth_face_detection_view.dart';
+export 'src/controller/youth_selfie_controller.dart';
 export 'src/biomarker_services/binah_sdk/view/youth_video_view.dart';
+export 'src/biomarker_services/binah_sdk/view/youth_face_detection_view.dart';
 export 'src/biomarker_services/liqa_sdk/youth_selfie_capturing.dart';
 export 'src/wrapper/enums.dart';
 export 'src/wrapper/youth_data_point.dart';