+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## [0.2.2]
+-  Fix `YouthAssets` path
+-  Fix and rename `YouthSelfieCaptureView`
+-  Add new biomarkers to `YouthBioScheme` according to Binah and Haut
+-  Add `YouthSelfieController`
+-  Add `ApiService`
+
 ## [0.2.1]
 -  Implemented Binah video service v2
 -  Add `YouthDataPoint` as main biomarker model
--  Fix Dinah dispose function 
+-  Fix Binah dispose function 
 
 ## [0.2.0]
 -  Implemented Binah video service v1